% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R, R/dijkstra_cpp.R
\name{dijkstra}
\alias{dijkstra}
\title{Implementation of dijkstra algorithm to find the shortest distance.}
\usage{
dijkstra(graph, init_node)

dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe with 3 variables v1, v2 and w ie edges of the graph from v1 to v2 with weight w.}

\item{init_node}{A number which is considered to be the initial point.}
}
\value{
The shortest distance from \code{init_node} to all other nodes in the graph.

The shortest distance from \code{init_node} to all other nodes in the graph.
}
\description{
This algorithm calculates the shortest distance between initial node and all the other nodes and saves the result into a vector.

This algorithm calculates the shortest distance between initial node and all the other nodes and saves the result into a vector.
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm

https://en.wikipedia.org/wiki/Graph#Mathematics

https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm

https://en.wikipedia.org/wiki/Graph#Mathematics
}
