% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean of two numbers.}
\usage{
euclidean(digit1, digit2)
}
\arguments{
\item{digit1}{A number.}

\item{digit2}{A number.}
}
\value{
The euclidean of \code{digit1} and \code{digit2}.
}
\description{
Implementation of Euclidean algorithm to find the greatest common divisor of two numbers.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
