% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Data containinng nodes and weights of graph}
\format{
A data frame with 54 rows and 3 variables:
\describe{
  \item{v1}{edges of the graph-from}
  \item{v2}{edges of the graph-to}
  \item{w}{weight of the edge from v1 to v2}
  ...
}
}
\source{
\url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\usage{
wiki_graph
}
\description{
A dataset is containing the edges and weights of graph.
}
\keyword{datasets}
